+++ src/Promotion/Changelog.md
@@ -2,6 +2,12 @@
 
 ## 5.x Series
 
+## Unreleased
+##### 2025-XX-YY
+
+- Added the cart item percent discount promotion action type
+- Changed the cart fixed and percent discount promotion action types to only act on carts, not on cart items
+
 ## 5.0.0
 ##### 2025-09-03